+++ CHANGELOG.md
@@ -7,10 +7,12 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 1.1.11
 
-2024 Jul 3 [in progress]
+2024 Jul 15 [in progress]
 
 * Removed translations for developer tools.
 * Fixed USB Inrush analysis for paths with spaces  #278
+* Improved German translation.
+* Fixed StatisticsRecordConfigDialog title.
 
 
 ## 1.1.10